[packaging] postinst: always enable the phosh service
Do some hoops and use deb-systemd-helper manually. Usually it's added
automatically by debhelper, but since we don't ship the service here
we need to do that automatically.

The phosh service is special so it makes sense that the upstream packaging
do not enable it by default. If one is going to have this package installed,
though, we can reasonably infer that the phosh being enabled by default is
the expected outcome.

(pre|post)rm aren't handled on purpose.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>